# [receiver/vcenter] Collection Time Performance Enhancement