… .chloggen/feat_vcenter-performance-tweaks.yaml
… receiver/vcenterreceiver/client_test.go
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-datastore.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-host-name.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-host-perf-counters.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-host-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-host.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-resource-pool-children.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-resource-pool-owner.xml
… receiver/vcenterreceiver/internal/mockserver/responses/cluster-resource-pool-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-datastore.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-default-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-host-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-host.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-resource-pool-owner.xml
… receiver/vcenterreceiver/internal/mockserver/responses/compute-resource-pool-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/datacenter-list.xml
… receiver/vcenterreceiver/internal/mockserver/responses/datastore-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/host-default-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/host-performance-counters.xml
… receiver/vcenterreceiver/internal/mockserver/responses/resource-pool-default-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/virtual-app-children.xml
… receiver/vcenterreceiver/internal/mockserver/responses/virtual-app-owner.xml
… receiver/vcenterreceiver/internal/mockserver/responses/virtual-app-properties.xml
… receiver/vcenterreceiver/internal/mockserver/responses/vm-default-properties.xml
… receiver/vcenterreceiver/metrics.go
… receiver/vcenterreceiver/resources.go
… receiver/vcenterreceiver/scraper.go