**Description:** <Describe what has changed.>
There were already some improvements made as far as how networks calls were made centered around Virtual Machines. This allowed collection times to decrease from ~90s to ~27s in an environment with 1 Cluster, 2 Hosts, & 280 VMs.

Making similar changes for all resource types helped to further decrease collection times. Now collection time has decreased from ~27s to <~3s for the same environment.

Here's a general list of the changes made:
- Now makes all network calls (per datacenter) first and stores returned data.
- Processes this data afterwards to convert to OTEL resources/metrics (refactored to new file).
- Moves all metric recording to metrics.go to keep consistent.
- Moves all resource builder creation to resources.go to keep consistent.
- Updates/fixes tests.

**Link to tracking Issue:** <Issue number if applicable>
#31837 Although this issue prescribes a solution to the problem (goroutines) which ended up not being necessary

**Testing:** <Describe what testing was performed and which tests were added.>
Unit Tests & Integration Tests Passing as well as Manual Testing in Local Environments

**Documentation:** <Describe the documentation added.>
N/A