@djaglowski 2024-05-13
> Updates/fixes tests.

If this is purely a performance enhancement, we should have no changes to tests.

Moreover, such extensive refactoring when accompanied by extensive test changes make it very difficult to establish that the changes are correct.