@StefanKurek 2024-05-13
> > Updates/fixes tests.
> 
> If this is purely a performance enhancement, we should have no changes to tests.
> 
> Moreover, such extensive refactoring when accompanied by extensive test changes make it very difficult to establish that the changes are correct.

@djaglowski Yeah, it's a tough one. In this refactor client makes a completely different set of calls with this refactor which means the most of the mock server's responses are outdated.

What exists in main today (outside of VMs which we have already updated), is that the client makes a call to get a list of resources (hosts, resource pools, whatever). What is returned is a completely bare bones version of these resources with basically a "path", a "reference" and a "client" to make more calls to get any properties. Then in the code as we want any properties for any individual resource, behind the scene the internal client makes an additional network call to grab that property or list of properties.

The whole idea behind this refactor is to make a different kind of request/response which grabs both the resources along with whatever properties we want up front (and not just get these skeletons back).